pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "mach2"
version = "0.0.2"
authors = [
  { name="Mrinmoy Saha Roddur", email="mroddur2@illinois.edu" },
]
description = "Migration Analysis of Clonal History 2"
readme = "README.md"
requires-python = ">=3.7"
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Bio-Informatics"
]
keywords = ["Computational biology", "Bioinformatics", "Cancer", "Metastasis", "Phylogeny"]
dependencies = [
    'numpy',
    'pandas',
    'pyomo',
    'networkx',
    'gurobi'
]

[project.optional-dependencies]
jupyter = [
    "jupyterlab",
    "python-graphviz",
]

[project.scripts]
mach2 = "mach2.mach:main"

[project.urls]
"Homepage" = "https://github.com/elkebir-group/MACH2"
"Bug Tracker" = "https://github.com/elkebir-group/MACH2/issues"